refactor: change connection error to be yellow
The connection error component showed a red danger as if something went
very wrong. This should be reserved if something goes wrong that halts
everything, such as the app crashing or permission errors. With generic
connection errors we should probably use a warning that something went
wrong with the connection itself but it can be refreshed

In this case, we change the icon to be a triangle instead of an
exclamation and also change color to be yellow instead of red
