+++ frontend/src/components/results/copr.js
@@ -9,7 +9,7 @@ import {
     Title,
 } from "@patternfly/react-core";
 
-import ConnectionError from "../error";
+import ConnectionError from "../error_connection";
 import Preloader from "../preloader";
 import TriggerLink from "../trigger_link";
 import { StatusLabel } from "../status_labels";
